phase-b2: keystone wave E-T + phase-c link setup (101 files; cascade-reconciled to green)

+++ src/runtime/server/InspectorBunFrontendDevServerAgent.rs
@@ -180,7 +180,7 @@ impl BunFrontendDevServerAgent {
                     connection_id,
                     url,
                     match route_bundle_id {
-                        Some(id) => id.get(),
+                        Some(id) => id.get() as i32,
                         None => -1,
                     },
                 );
